project: rename Sepulchre
project: rename Sepulchre to Hallowed Sepulchre
