+++ settings.gradle.kts
@@ -39,6 +39,7 @@ include(":entityhiderextended")
 include(":fightcave")
 include(":gauntlet")
 include(":grotesqueguardians")
+include(":hallowedsepulchre")
 include(":hideprayers")
 include(":hideunder")
 include(":hydra")
@@ -55,7 +56,6 @@ include(":pvptools")
 include(":reorderprayers")
 include(":runedoku")
 include(":socket")
-include(":sepulchre")
 include(":specbar")
 include("specialcounterextended")
 include(":spellbook")
